Construct DOM for the REPL
One objective is to have the REPL bootstrap itself by creating it's own
DOM and binding behaviour to it (rather than presuming some existing DOM
will be present).

In this commit, a very basic little DOM construction library is
introduced. It's purpose is to take a simple JavaScript data structure
representing the DOM, build the appropriate DOM and mutate the structure
with references to the constructed DOM - and give it back. This should
provide a way to access DOM without using the expensive DOM API - and
modify our DOM in a declaritive fashion, should we wish.

Additionally, some minimal structure representing the REPL is declared,
and a set of lenses that will provide access to interesting parts of the
DOM when applied to the structure.